Bug fixing startup process

+++ dpportal.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/bash -x
 
 ### Wrapper for DPPortal runtime ###
 
@@ -12,6 +12,11 @@ q="\e[1;32m"         # questions (green)
 ## Runtime directory
 runtimeDir=/etc/dpportal/run
 
+scripDir="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+
+theMainArg="$1"
+
+echo "$theMainArg"
 
 ###Some more colour and formatting vars, some may become redundant as I continue to dev and tidy up my project
 _bold=$(tput bold)
@@ -58,61 +63,59 @@ function processArgs()
 {
 
 	if [[ $# -gt 1 ]]; then
-    	clear
-        _printPoweredBy
-    	_error "Only one arguement is required."
-        _die "I'm off bruv!"
+		clear
+		_printPoweredBy
+		_error "Only one arguement is required."
+		_die "I'm off bruv!"
 	fi
 
 
-	argCommand=$1
-	
+	argCommand="$theMainArg"
+
     # Get function called and refer shell to relavent script
-    for arg in "$@"
-    do
-        case $arg in
-            setbrand)
-				cd $runtimeDir
-				./setBrand.sh
-            ;;
-            run)
-				cd $runtimeDir
-				./bootstrap.sh
-			;;
-            autorun)
-            	cd $runtimeDir
-                ./bootstrap.sh 'ar'
-            ;;
-            stop)
-            	stopBerate
-			;;
-			uninstall)
-				read -p "Are you sure you want to uninstall this fine tool?! (yes/no)" uninVar
-                if [[ "$uninVar" = "yes" ]]; then
-                	cd $runtimeDir
-                	./uninstall.sh
-				elif [[ "$uninvar" = "no" ]]; then
-					clear
-                    _printPoweredBy
-                    echo -e "$info\n\nYou said no, DPPortal will now exit. Bye..!"
-                    sleep 3
-                    ctrl_c
-				else
-                    clear
-                    _printPoweredBy
-                    echo -e "$info\n\nYour answer didn't make any sense. I'm dipping bruv!\n"
-                    sleep 3
-                    ctrl_c
-				fi
-            ;;
-            -h|--help)
-                printUsage()
-            ;;
-            *)
-                printUsage()
-            ;;
+
+	case "$theMainArg" in
+	setbrand)
+		"$scriptDir/setBrand.sh"
+	;;
+		run)
+			"$runtimeDir/bootstrap.sh"
+	;;
+	autorun)
+		"$runtimeDir/bootstrap.sh ar"
+	;;
+	stop)
+		stopBerate
+	;;
+	uninstall)
+		read -p "Are you sure you want to uninstall this fine tool?! (yes/no)" uninVar
+		if [[ "$uninVar" = "yes" ]]; then
+			cd $runtimeDir
+			./uninstall.sh
+		elif [[ "$uninvar" = "no" ]]; then
+			clear
+			_printPoweredBy
+			echo -e "$info\n\nYou said no, DPPortal will now exit. Bye..!"
+			sleep 3
+			ctrl_c
+		else
+			clear
+			_printPoweredBy
+			echo -e "$info\n\nYour answer didn't make any sense. I'm dipping bruv!\n"
+			sleep 3
+			ctrl_c
+		fi
+	;;
+	help)
+		printUsage
+	;;
+	"")
+		"$runtimeDir/bootstrap.sh"
+	;;
+	*)
+		printUsage
+	;;
         esac
-    done
 
 }
 
@@ -122,12 +125,56 @@ function stopBerate()
 	clear
 	_printPoweredBy
 	beratePID=$(pgrep berate_ap | head -1)
-	kill -9 $beratePID
-    a2dissite dpportal.conf
-    a2dissite dppdisplay.conf
-    systemctl stop apache2
-	systemctl stop mysql
-    echo -e "$info\nYou have chosen to quit DPPortal and the associated processes have been stopped."
+	#kill -9 $beratePID
+	#a2dissite dpportal.conf
+	#a2dissite dppdisplay.conf
+	#systemctl stop apache2
+	#systemctl stop mysql
+	#echo -e "$info\nYou have chosen to quit DPPortal and the associated processes have been stopped."
+	
+	
+	echo "Shutting down services safely..."
+ 
+    # Stop apache2 if running
+    if systemctl is-active --quiet apache2; then
+        echo "Stopping apache2..."
+        systemctl stop apache2
+    else
+        echo "apache2 is not running."
+    fi
+ 
+    # Stop mysql if running
+    if systemctl is-active --quiet mysql; then
+        echo "Stopping mysql..."
+        systemctl stop mysql
+    else
+        echo "mysql is not running."
+    fi
+ 
+    # Disable apache sites if enabled
+    if [ -e /etc/apache2/sites-enabled/dpportal.conf ]; then
+        echo "Disabling the captive portal site."
+        a2dissite dpportal.conf
+    else
+        echo "dpportal.conf is not enabled."
+    fi
+ 
+    if [ -e /etc/apache2/sites-enabled/dppdisplay.conf ]; then
+        echo "Disabling the loot display site."
+        a2dissite dppdisplay.conf
+    else
+        echo "dppdisplay.conf is not enabled."
+    fi
+ 
+    # Kill berate_ap process if running
+    if pgrep -x "berate_ap" > /dev/null; then
+        echo "Killing your filthy credential thieving access point"
+        pkill -9 "berate_ap"
+    else
+        echo "berate_ap is not running."
+    fi
+ 
+    echo "You're done pal. Slan leat a chara!"
 
 }
